This commit includes changes to the NPC system including completion of the underlying NPC code. Additional changes: - Created first NPC type - Created NPC collision handler - Created NPC spawning scheduler - Code improvements to physics filter classes - Created event/listener system for certain world changes (for now just entity added/removed from world) - NPCs add players points on death - Collision with NPC causes player damage - NPC Collision with bullets causes NPC damage and destroys the bullet.
Only TODO: Send initial NPC list to client upon joining game

+++ game-backend/src/main/java/com/becky/networking/message/NpcInfo.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class NpcInfo implements NetworkedMessage {
     private String type; //the classname of the npc type. Doesn't need to be set if state != NPC_STATE_NEW
-    private int npcId;
+    private long npcId;
     private float positionX;
     private float positionY;
     private float angle;
@@ -18,7 +18,7 @@ public class NpcInfo implements NetworkedMessage {
         this.type = type;
     }
 
-    public void setNpcId(final int npcId) {
+    public void setNpcId(final long npcId) {
         this.npcId = npcId;
     }
 
@@ -46,7 +46,7 @@ public class NpcInfo implements NetworkedMessage {
         return this.type;
     }
 
-    public int getNpcId() {
+    public long getNpcId() {
         return this.npcId;
     }
